manifest: move subscription-related functions
Move subscription-related functions from manifest/os.go to
manifest/subscription.go.  This is an intra-package move so it has no
functional effect.
